refeactor: create Interceptor.java and move git token & claim code
